server update: read desc for changelog
Impl collision based battles. Battles are still rnd determined and
instant.
Impl sieging for castles, based on unit dmg values (see below).
Impl unit dmg values, disagree with them? feel free to contact me abt
it.
Impl a non-timer based start up. Game will start once a certain number
of people join (set at 4 atm).
Units now move according to their speed.
Castles now provide income based on their upgrade level, does not cap at
the moment.
Income is now set on the number of castles you have and the upgrade
levels of those castles.
Game now checks to see if a buy order for a unit is valid (no more send
ws cmds to buy a unit in the middle of the field).
Added castles in the middle of the field, not random at the moment (we
may add random placement or not).
Should have fixed synchronization issues with the for loop iterators,
haven't been able to replicate the previous issue, so its my best guess.

+++ CastleClashersServer/CastleClashersServer/src/com/threeC/beans/Player.java
@@ -10,7 +10,7 @@ public class Player implements JSONStringifiable {
 	
 	public String name;
 	public int gold = 25;
-	public int income = 0;
+	public float income = 0;
 	
 	public long alliance = -1l;
 	public long war = -1l;
